osd: allow for injecting extra environment variables
This patch is both stripped down (no support for extra argvs)
and extended (envs at both prepare and main stages) version
of the Travis Nielsen's patch [1] to accommodate crimson-osd
in Rook. Also, the naming convention isn't crimson-specific
anymore as the mechanism could be useful with the classical
OSD as well (e.g. tuning TCMalloc or even the dynamic linker).

The need for the patch comes from the fact that, although
crimson finally exposes the same CLI interface as the classical
OSD, the broadly used development builds have ASan built in.
As ASan, by default, complains if it isn't the very first loaded
DSO, we need to set the `ASAN_OPTIONS` environment variable to
`verify_asan_link_order=0` to mitigate the early aborts.

Signed-off-by: Radoslaw Zarzynski <rzarzyns@redhat.com>

[1]: https://github.com/rook/rook/pull/8763