pool: fix possible segfault in remote executor
Original error check was:
if stderr != "" || err != nil {
	err = errors.Errorf("%s. %s", err.Error(), stderr)
}

Ceph uses stderr to output debug info, not just error messages. So
checking for stderr != "" as a condition for error could have errors. When
there is no error but there is stderr text, this code will segfault when
accessing err.Error().

Fix the check to be the usual Go error check as below:
if err != nil {...}

Additionally, output a warning message in the log containing the stderr
text to allow the Rook team to debug further in the future.

Signed-off-by: Blaine Gardner <blaine.gardner@redhat.com>

[todo: squash] treat stdout as error