core: retry on update conflict
The API server will return `Conflict` when we try to update an outdated
resource. This can happen in scenarios where we try to update the status
but the CR object spec has been updated by the user.A
We will retry we back off 5 times. We were using this mechanism already
but in a single controller only. Now we generalize this behaviour to all
controllers.
We finally get rid of all log messages like:

"the object has been modified
please apply your changes to the latest version and try again"

Signed-off-by: Sébastien Han <seb@redhat.com>