osd: remove osd with purge instead of destroy
The osd destroy command leaves the osd id in use instead
of fully purging the osd id. If the osd id is not also
removed, it gives the impression that there is still
something that needs to be cleaned up from the old osd.

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>
(cherry picked from commit 879bde4416aeb85b4bf1958236be342c27dd335b)