docs: remove obsolete section of upgrade guide
The bluestore_fsck_quick_fix_on_mount fix was only needed when upgrading to
versions of Ceph older than v16.2.6. Everyone on pacific should now be
on v16.2.7, so we can remove this section from the v1.9 upgrade guide.

Also cleaning up a few other small issues in the guide.

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>