helm: correctly check external cluster for prometheus rules
When the prometheus rules are created, if external settings
were not specified, the helm chart would fail due to a nil ref
in the external settings. Now we check both for the existence
of external settings and whether the external settings are enabled.

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>
(cherry picked from commit 257dc61391736ced87b855c784e6ec54ba52075b)