docs: consistent toc headers for crds
The titles for the CRD documentation pages in the table
of contents were not consistent. The full name for the CRD
can be included in the full doc page, but in the TOC
we abbreviate the names for simplicity and readability.

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>
(cherry picked from commit a71e361cef90da49aa381cbab5dcfc244a374e1d)