build: add a placeholder rgw role for csv generation
The csv will not properly include the rook-ceph-rgw service
account unless there is a role and binding that references
the service account. No roles are needed yet for the rgw
daemon, so this will add a placeholder only for the purposes
of csv generation.

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>
(cherry picked from commit e1550e7551019efb2dc941f0a7ed00267b221389)