monitoring: disable new alerts that are not applicable to rook
The new alerts picked up for Rook v1.9 directly from the ceph repo
were not all rook-compliant. The alert for the prometheus job not
running is not needed since Rook and K8s go to great lengths to
keep the mgr pod running already.

The alert for the cluster filling up soon is also disabled
until we find how to get the formula working in rook clusters.

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>