csi: rework how to save csi configmap details
The rados namespace and subvolume groups were being saved to the
csi configmap in their own entries, but their mons were not being
updated since rook can't tell which cluster they belonged to.
Now the rados namespaces and subvolume groups are added within
the same cluster entries so the mons only need to be updated in
a single place whenever there is an update.

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>