core: rgw realm pull spec is optional
The endpoint on the rgw realm pull spec is causing a failure
if not specified even when it is not expected to be
required. Now the endpoint has the json property to be
ommitted if not specified.

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>
(cherry picked from commit 7f709776b5017909f38d5fe47bc5cebb4e40530b)