object: handle hpa scenarios for object store reconcilation
If Horizontal Pod Scalers is configured scaled up replica count for ceph
object store, then during reconcilation Rook Operator will reset the
value for replica and suddenly HPA set it back. To avoid this sceneraios
reconcilation check whether HPA is configured or not and set the replica
value according to that.

Signed-off-by: Jiffin Tony Thottan <thottanjiffin@gmail.com>