operator: increase memory limit of operator
The operator is very bursty and when starting up may need more memory
to reconcile all the CRs initially before it calms down.
The default of 256Mi was not sufficient for many clusters, so now
we bump up the limit to 512Mi.

This is a follow-up from the same change to the helm chart in
https://github.com/rook/rook/pull/10195.

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>
(cherry picked from commit 60b72b6d48a7f403316fb814c1fe8121772ccbdc)