docs: add prometheus doc link to pick up rook service monitor
If prometheus is already installed in a cluster, it may not be configured
to watch for third-party service monitors such as for Rook.
To help with this configuration, a link is added to the doc for
enabling the third-party service monitors.

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>
(cherry picked from commit 91a8011a614ad3eb1fcffcdf06c413e3aa363bc8)