core: append filesystem property on disk
When detecting the device property the filesystem was not passed but
used later to validate if the device should be taken or not. Now we read
the filesystem info from "lsblk" and populate it in the device type.

Closes: https://github.com/rook/rook/issues/9470
Signed-off-by: Sébastien Han <seb@redhat.com>
(cherry picked from commit 8c1bf0f3f0460e5fcb699d5c911978330cdc8729)