build: disable docs build in 1.8 branch
With the full rework of the docs framework, we cannot support automated
updating of docs from branches where the doc framework has not been
backported. We are largely done with the 1.8 releases, so we can just
disable publishing the docs for the 1.8 branch. If further updates
are required in the docs, we can update the legacy docs directly in
the rook.github.io repo.

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>