manifest: inline toolbox script for direct use of ceph image
The toolbox is really only used for ceph commands. The main reason
the rook image was being used in the toolbox pod was for the script
that generates the ceph.conf and updates it whenever the mons are
updated during mon failover.

Now the ceph image can be specified directly by moving the script
inline with the container definition instead of being required
in the image.

The rook image will still contain the script for backward compatibility
and for scenarios where the rook binary may still be needed

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>