monitoring: fix pool growth warning grouping
Predict linear will continue to predict on exporter data from old pods.
We can take the max value for every hour in the 2 day window before we
linearly predict to get the most conservative estimate. This will only
affect the hour in which a new mgr pod is started.

Fixes #10691

Signed-off-by: Jamison Lofthouse <jamison.lofthouse@gmail.com>