core: admin test cluster info should default to rookoperator key
The AdminClusterInfo method is called in numerous places for a quick
placeholder for the cluster info, and makes assumptions about the username,
but doesn't set other fields such as the user secret. Now we can assume
the cluster info is based on the rookoperator keyring by default.
In the upgrade test we just need to revert to the client.admin
keyring before the upgrade.

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>