core: wait for rookoperator keyring in most controllers
If a converged cluster, we expect all of the controllers to use the
rookoperator keyring. The exceptions are that external clusters will
use any keyring available, the cluster controller of course will use
any keyring as it owns generating the rookoperator keyring, and the
csi driver will use any keyring.

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>