csi: remove findCSIChange from predicate
findCSIChange will work only for single
line values/configration in the configmap,
If the value is multiline the regex will not
be able to indentify the change because of this
one no configurations will be applied to the
csi driver. Removing this check so that if there
is any update on the configmap we will
reconcile the csi driver.

Note:- This might looks heavy but updating of config
map is not a frequent event.

Signed-off-by: Madhu Rajanna <madhupr007@gmail.com>