docs: warn to upgrade to 1.9.10 before k8s 1.25
The helm chart is blocked from upgrading to newer versions of Rook
if v1.9.10 is not installed before upgrading to K8s 1.25. The
helm chart is confused by the PSPs that still exist in the installed
chart and fails to upgrade to the new chart.

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>
(cherry picked from commit 9cad930b4d57cff92ba925762f0d4cf00fc50532)