osd: prepare job needs significant more memory for provisioning
The OSD creation may need to burst during OSD provisioning depending
on the size of the OSD or similar factors. If the OSD prepare job is
OOM killed it will cause OSD provisioning to fail and have various
side effects that are difficult to troubleshoot to get the OSD
to succeed. So we increase the recommendation significantly to avoid
the OOM kill.

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>
(cherry picked from commit 92c52c4a2eb9b380f536c868b556bb3d3482db88)